update project godot with opencl from file and from string

+++ project/FindMind.cl
@@ -1,4 +1,4 @@
-kernel void func(ulong n, global const int * data, global int * res){
+kernel void func(ulong n, global const float * data, global float * res){
 	size_t i = get_global_id(0);
-	res[i] = data[i] * data[i];
-}
+ 	res[i] = data[i] * data[i];
+ }
